move Scopes to another file
